requirements.txt
streamlit
langchain-community
langchain
pypdf
pdfminer.six